--- jaseci_core/jaseci/svc/action_optimizer/configs/__init__.py
@@ -1,33 +1,0 @@
-from jaseci.svc.action_optimizer.configs.bi_enc_action_config import (
-    BI_ENC_ACTION_CONFIG,
-)
-from jaseci.svc.action_optimizer.configs.cl_summer_action_config import (
-    CL_SUMMER_ACTION_CONFIG,
-)
-from jaseci.svc.action_optimizer.configs.ent_ext_action_config import (
-    ENT_EXT_ACTION_CONFIG,
-)
-from jaseci.svc.action_optimizer.configs.fast_enc_action_config import (
-    FAST_ENC_ACTION_CONFIG,
-)
-from jaseci.svc.action_optimizer.configs.pdf_ext_action_config import (
-    PDF_EXT_ACTION_CONFIG,
-)
-from jaseci.svc.action_optimizer.configs.text_seg_action_config import (
-    TEXT_SEG_ACTION_CONFIG,
-)
-from jaseci.svc.action_optimizer.configs.tfm_ner_action_config import (
-    TFM_NER_ACTION_CONFIG,
-)
-from jaseci.svc.action_optimizer.configs.use_enc_action_config import (
-    USE_ENC_ACTION_CONFIG,
-)
-from jaseci.svc.action_optimizer.configs.use_qa_action_config import (
-    USE_QA_ACTION_CONFIG,
-)
-
-ACTION_CONFIGS = {
-    "use_enc": USE_ENC_ACTION_CONFIG,
-    "use_qa": USE_QA_ACTION_CONFIG,
-    "bi_enc": BI_ENC_ACTION_CONFIG,
-}
