+++ jaseci_core/jaseci/svc/actions_optimizer/configs/text_seg_action_config.py
