+++ jaseci_core/jaseci/svc/actions_optimizer/configs/use_enc_action_config.py
@@ -1,15 +1,15 @@
-USE_QA_ACTION_CONFIG = {
-    "module": "jaseci_ai_kit.use_qa",
+USE_ENC_ACTION_CONFIG = {
+    "module": "jaseci_ai_kit.use_enc",
     "remote": {
         "Service": {
             "kind": "Service",
             "apiVersion": "v1",
-            "metadata": {"name": "jaseci-use-qa", "creationTimestamp": None},
+            "metadata": {"name": "jaseci-use-enc", "creationTimestamp": None},
             "spec": {
                 "ports": [
                     {"name": "http", "protocol": "TCP", "port": 80, "targetPort": 80}
                 ],
-                "selector": {"pod": "jaseci-use-qa"},
+                "selector": {"pod": "jaseci-use-enc"},
                 "type": "ClusterIP",
                 "sessionAffinity": "None",
                 "internalTrafficPolicy": "Cluster",
@@ -20,40 +20,40 @@ USE_QA_ACTION_CONFIG = {
             "kind": "ConfigMap",
             "apiVersion": "v1",
             "metadata": {
-                "name": "jaseci-use-qa-up",
+                "name": "jaseci-use-enc-up",
                 "namespace": "default",
                 "creationTimestamp": None,
             },
             "data": {
-                "prod_up": "pip install jaseci-ai-kit==$1.3.5.22\nuvicorn jaseci_ai_kit.use_qa:serv_actions --host 0.0.0.0 --port 80"
+                "prod_up": "pip install jaseci-ai-kit==1.3.5.22\nuvicorn jaseci_ai_kit.use_enc:serv_actions --host 0.0.0.0 --port 80 --workers 4"
             },
         },
         "Deployment": {
             "kind": "Deployment",
             "apiVersion": "apps/v1",
-            "metadata": {"name": "jaseci-use-qa", "creationTimestamp": None},
+            "metadata": {"name": "jaseci-use-enc", "creationTimestamp": None},
             "spec": {
                 "replicas": 1,
-                "selector": {"matchLabels": {"pod": "jaseci-use-qa"}},
+                "selector": {"matchLabels": {"pod": "jaseci-use-enc"}},
                 "template": {
                     "metadata": {
-                        "name": "jaseci-use-qa",
+                        "name": "jaseci-use-enc",
                         "creationTimestamp": None,
-                        "labels": {"pod": "jaseci-use-qa"},
+                        "labels": {"pod": "jaseci-use-enc"},
                     },
                     "spec": {
                         "volumes": [
                             {
                                 "name": "prod-script",
                                 "configMap": {
-                                    "name": "jaseci-use-qa-up",
+                                    "name": "jaseci-use-enc-up",
                                     "defaultMode": 420,
                                 },
                             }
                         ],
                         "containers": [
                             {
-                                "name": "jaseci-use-qa",
+                                "name": "jaseci-use-enc",
                                 "image": "jaseci/jaseci-ai-kit:latest",
                                 "command": ["bash", "-c", "source script/prod_up"],
                                 "ports": [{"containerPort": 80, "protocol": "TCP"}],
